refactor: move component tests into own dir
