+++ src/web/src/__tests__/components/WSEditorClientConfig.test.tsx
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { render } from "./test-utils";
-import WSEditorClientConfigDialog from "../views/workspace/WSEditorClientConfig";
-import { workspaceApi, specsApi, errorHandlerApi } from "../services";
+import { render } from "../test-utils";
+import WSEditorClientConfigDialog from "../../views/workspace/WSEditorClientConfig";
+import { workspaceApi, specsApi, errorHandlerApi } from "../../services";
 
-vi.mock("../services", () => ({
+vi.mock("../../services", () => ({
   workspaceApi: {
     getClientConfig: vi.fn(),
     updateClientConfig: vi.fn(),
